Rename SubPlot -> Plot.
File names now match class names.
